backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
pydantic==2.13.1
python-dotenv==1.2.2
supabase==2.28.3
groq==1.1.2
httpx==0.28.1
requests==2.33.1
